Handle Forbidden And Update Controllers Authorization Roles

+++ DepiFinalProject/Controllers/FlashsalesController.cs
@@ -7,6 +7,7 @@ namespace DepiFinalProject.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
+    [Authorize]
     public class FlashSalesController : ControllerBase
     {
         private readonly IFlashSaleService _service;
@@ -18,6 +19,11 @@ namespace DepiFinalProject.Controllers
 
         // GET: api/flashsales
         [HttpGet]
+        [AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> GetAll()
         {
             var flashSales = await _service.GetAllAsync();
@@ -26,6 +32,11 @@ namespace DepiFinalProject.Controllers
 
         // GET: api/flashsales/{id}
         [HttpGet("{id}")]
+        [AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> Get(int id)
         {
             var flashSale = await _service.GetByIdAsync(id);
@@ -38,9 +49,17 @@ namespace DepiFinalProject.Controllers
 
         // POST: api/flashsales
         [HttpPost]
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> Create([FromBody] CreateFlashSaleDto dto)
         {
+            if (!User.IsInRole("admin") )
+            {
+                return StatusCode(403, new { Error = "Only Allowed To Admin" });
+            }
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
@@ -54,9 +73,18 @@ namespace DepiFinalProject.Controllers
 
         // PUT: api/flashsales/{id}
         [HttpPut("{id}")]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> Update(int id, [FromBody] UpdateFlashSaleDto dto)
         {
+            if (!User.IsInRole("admin"))
+            {
+                return StatusCode(403, new { Error = "Only Allowed To Admin" });
+            }
+
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
@@ -70,9 +98,18 @@ namespace DepiFinalProject.Controllers
 
         // DELETE: api/flashsales/{id}
         [HttpDelete("{id}")]
-        [Authorize(Roles = "Admin")]
+        [Authorize(Roles = "admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!User.IsInRole("admin"))
+            {
+                return StatusCode(403, new { Error = "Only Allowed To Admin" });
+            }
+
             var result = await _service.DeleteAsync(id);
 
             if (!result)
